+++ packages/embed/src/initializers/build-options-from-attributes.ts
@@ -30,5 +30,7 @@ export const buildOptionsFromAttributes = (element: HTMLElement) => {
     autoClose: 'integerOrBoolean',
     shareGaInstance: 'boolean',
     forceTouch: 'boolean',
+    enableFullscreen: 'boolean',
+    inlineOnMobile: 'boolean',
   })
 }